refactor: BridgeGameStarter 패키지 변경
