+++ src/main/java/bridge/output/BridgeGameStarter.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.output;
 
 public class BridgeGameStarter {
     private static final String TEXT_GAME_START = "다리 건너기 게임을 시작합니다.";
